Implement handleArbitraryText method in TelegramBot; remove legacy handle_arbitraty_text.go file
