+++ internal/bot/bot.go
@@ -7,8 +7,10 @@ import (
 	"fmt"
 	"log/slog"
 
+	"github.com/2Cheetah/MedGuardianBot/internal/domain"
 	"github.com/2Cheetah/MedGuardianBot/internal/service"
 	"github.com/go-telegram/bot"
+	"github.com/go-telegram/bot/models"
 )
 
 type TelegramBot struct {
@@ -57,4 +59,21 @@ func (tb *TelegramBot) RegisterHandlerExactMatch(pattern string, h bot.HandlerFu
 		h,
 	)
 	slog.Debug("registered handler", "pattern", pattern, "id", id)
+}
+
+func (tb *TelegramBot) handleArbitraryText(ctx context.Context, b *bot.Bot, update *models.Update) {
+	userID := update.Message.From.ID
+	context := update.Message.Text
+	dialog := &domain.Dialog{
+		UserID:  userID,
+		Context: context,
+	}
+	msg, err := tb.DialogService.HandleDialog(dialog)
+	if err != nil {
+		slog.Error("couldn't handleArbitraryText", "error", err)
+		msg := fmt.Sprintf("Error while handling text:\n%s", update.Message.Text)
+		sendMsg(ctx, b, update.Message.Chat.ID, msg)
+		return
+	}
+	sendMsg(ctx, b, update.Message.Chat.ID, msg)
 }
